requirements.txt
numpy
opencv-python
scikit-learn
scikit-image
pillow
matplotlib